+++ eslint.config.mjs
@@ -2,10 +2,12 @@ import globals from "globals";
 import pluginJs from "@eslint/js";
 import tseslint from "typescript-eslint";
 import pluginVue from "eslint-plugin-vue";
+import eslintPluginUnicorn from 'eslint-plugin-unicorn';
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
   { files: ["**/*.{js,mjs,cjs,ts,vue}"] },
+  eslintPluginUnicorn.configs['flat/recommended'],
   { languageOptions: { globals: globals.browser } },
   pluginJs.configs.recommended,
   ...tseslint.configs.recommended,
@@ -15,8 +17,9 @@ export default [
     languageOptions: { parserOptions: { parser: tseslint.parser } },
   },
   {
-    "rules": {
-      "vue/no-undef-properties": "error"
-    }
-  }
+    rules: {
+      "vue/no-undef-properties": "error",
+      "unicorn/no-null": "off",
+    },
+  },
 ];